requirements.txt
altgraph==0.17.4
attrs==24.2.0
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
comtypes==1.4.11
et_xmlfile==2.0.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.6
lxml==5.4.0
MarkupSafe==3.0.2
MouseInfo==0.1.3
mss==10.0.0
mysql-connector-python==9.3.0
numpy==2.1.3
opencv-python==4.10.0.84
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pefile==2023.2.7
pillow==11.0.0
playwright==1.49.1
pluggy==1.5.0
psutil==7.0.0
PyAutoGUI==0.9.54
pycparser==2.22
pyee==12.0.0
pygame==2.6.1
PyGetWindow==0.0.9
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2025.0
PyMsgBox==1.0.9
pypandoc==1.15
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytesseract==0.3.13
pytest==8.3.4
pytest-base-url==2.1.0
pytest-html==4.1.1
pytest-metadata==3.1.1
pytest-playwright==0.6.2
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-slugify==8.0.4
pytube==15.0.0
pytubefix==8.12.2
pytweening==1.2.0
pytz==2025.2
pywin32-ctypes==0.2.3
requests==2.32.3
selenium==4.26.1
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tabulate==0.9.0
text-unidecode==1.3
trio==0.27.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.3
websocket-client==1.8.0
wsproto==1.2.0
yt-dlp==2025.2.19
